% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_size_estimator.R
\name{rwalk}
\alias{rwalk}
\title{Graph and Network Sampler}
\usage{
rwalk(nodes, steps, subset)
}
\arguments{
\item{nodes}{Graph or network data, in the form of a table of two columns representing edges between pairs of nodes: from (column 1) and to (column 2).}

\item{steps}{Number of steps to take before the next node is sampled in the random walk.}

\item{subset}{The number of nodes to be sampled in the random walk.}
}
\description{
A short function to perform simple random walk sampling on graph or network data.
Requires the igraph package.
}
\details{
The output of the random walk can be accessed via two commands:
\itemize{
\item \code{graph.Random.Walk} returns a vector of nodes sampled;
\item \code{rw.Node.Degrees} returns a vector of the corresponding node degrees.
}
}
\examples{
# Performs a random walk on the graph or network data,
# sampling each consectuive node until 1000 nodes have been sampled:
walk(nodes, 1, 1000)

# Performs a random walk on the graph or network data,
# sampling nodes 3 steps apart until 5000 nodes have been sampled:
walk(nodes, 3, 5000)
}
\keyword{graph}
\keyword{network}
\keyword{random}
\keyword{sampling}
\keyword{walk}
