% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_size_estimator.R
\name{size.Est}
\alias{size.Est}
\title{Graph and Network Size Estimator}
\usage{
size.Est(node.File, degree.File, rw.Step, size, N, max.Rs.Division, g,
  wn.Constant)
}
\arguments{
\item{node.File}{Vector of nodes sampled from the full graph or network.}

\item{degree.File}{Vector of the corresponding degrees of the nodes in \code{node.File}.}

\item{rw.Step}{Number of steps to take before the next node is sampled in the simple random walk.}

\item{size}{Sample size increment.}

\item{N}{True population value.}

\item{max.Rs.Division}{Maximum random sample divisor.}

\item{g}{Srivastava and Bhatnagar's constant, \code{g = 1} gives \code{N_B = N_A}.}

\item{wn.Constant}{Wither's and Nadarajah's constant (A).}
}
\description{
A short function to compute estimates for the size of a graph or network, using small samples of data from the graph or network.
Requires the ggplot2 package.
}
\details{
Can be used in conjunction with \code{walk}, where \code{node.File = graph.Random.Walk},
and \code{degree.File = rw.Node.Degrees}.

\code{size} can be computed by \deqn{size = (length(node.File)/rw.Step)/(max.Rs.Division + 1)}
}
\examples{
# e.g. node.File and degree.File contain
# 2,000,000 nodes and degree values, respectively.
# The following computes estimates of
# the population size, and bias of the estimators, at sample sizes
# from n = 400 to 3600, in stpes of 400, with g = 10, and A = 0.5.
size.Est(node.File, degree.File, 500, 400, 42000000, 9, 10, 0.5)
}
\keyword{estimator}
\keyword{graph}
\keyword{network}
\keyword{size}
